tests/wpt/tests/css/css-shapes/shape-functions/polygon-function-round-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Shapes Module Level 1: parsing the polygon() function with the round modifier</title>
<link rel="help" href="https://drafts.csswg.org/css-shapes-1/#funcdef-basic-shape-polygon">
<meta name="assert" content="Tests parsing of the round modifier in the polygon() function">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value(
    "shape-outside",
    "polygon(round 0, 1px 2px, 3px 4px)",
    "polygon(1px 2px, 3px 4px)");
test_valid_value(
    "shape-outside",
    "polygon(round 0px, 1px 2px, 3px 4px)",
    "polygon(1px 2px, 3px 4px)");
test_valid_value("shape-outside", "polygon(round 1px, 1% 2%)");
test_valid_value(
    "shape-outside",
    "polygon(nonzero round 1px, 1px 2px, 3em 4em)",
    "polygon(round 1px, 1px 2px, 3em 4em)");
test_valid_value(
    "shape-outside",
    "polygon(evenodd round 1px, 1px 2px, 3em 4em, 5pt 6%)");
test_valid_value(
    "shape-outside",
    "polygon(round calc(1em + 2px), 1px 2px, 3px 4px)");
</script>
</body>
</html>
